condense module hooks pt.2
