+++ graphcap_studio/src/features/perspectives/hooks/index.ts
@@ -13,7 +13,7 @@ export { usePerspectives } from "./usePerspectives";
 export { usePerspectiveModules } from "./usePerspectiveModules";
 export { useGeneratePerspectiveCaption } from "./useGeneratePerspectiveCaption";
 export { useImagePerspectives } from "./useImagePerspectives";
-export { useModules } from "./useModules";
+
 
 // Utilities
 export * from "./utils";
